Minor bugfix in shiftout, start on refreshDisplay
Some variables were erroneously defined for shitfout, along with some
registers.

refreshDisplay-methods is now a bit more correct (it will output
somethings). However, the code needs more review before it is ready.

There is also a lot of flickering right now, which is not looking good...

+++ fancontroller.h
@@ -14,7 +14,7 @@
 	uint16_t readFanCurrent(uint8_t channel);;
 	void checkConnection(void);
 	void refreshDisplay(void);
-	void shiftout(uint8_t input);
+	void shiftout(uint32_t input);
 
 	//PORT LOCATIONS
 		//Shift registry
